fix: 컴포넌트 파일명 및 변수명 컨벤션에 맞춰 수정
