+++ src/Components/button/MediumSmallButton.jsx
@@ -1,23 +1,23 @@
 import styled from "styled-components";
 
-const Button = styled.button`
+const MediumSmallBtn = styled.button`
   width: 90px;
   padding: 7px 0;
   border-radius: 32px;
   border: none;
-  background-color: var(--accentColor);
+  background-color: var(--mainColor);
   font-family: inherit;
   font-weight: 500;
   font-size: 1.4rem;
   color: white;
 `;
 
-function MediumSmallDisableButton() {
+function MediumSmallButton() {
   return (
     <>
-      <Button>저장</Button>
+      <MediumSmallBtn>저장</MediumSmallBtn>
     </>
   );
 }
 
-export default MediumSmallDisableButton;
+export default MediumSmallButton;
