+++ src/Components/button/SmallButton.jsx
@@ -1,22 +1,23 @@
 import styled from "styled-components";
 
-const Button = styled.button`
+const SmallBtn = styled.button`
   width: 56px;
   padding: 7px 0;
   border-radius: 26px;
-  border: 1px solid #dbdbdb;
-  background-color: white;
+  border: none;
+  background-color: var(--mainColor);
   font-family: inherit;
   font-weight: 400;
   font-size: 1.2rem;
+  color: white;
 `;
 
-function SmallActivButton() {
+function SmallButton() {
   return (
     <>
-      <Button>취소</Button>
+      <SmallBtn>팔로우</SmallBtn>
     </>
   );
 }
 
-export default SmallActivButton;
+export default SmallButton;
